requirements.txt
pandas==2.2.2
python-dotenv==1.0.1
pipreqs==0.5.0
termcolor==2.4.0
setuptools==72.1.0
twine==5.1.1
wheel==0.44.0

